Wikimedia Event Utilities initial implementation
EventStreamConfig aides in requesting local or remote stream configs.
It is meant to be used with the EventStreamConfig MediaWiki
API endpoint.

EventStream represents a single event stream and provides methods
for looking up schema and topics and for
creating canary events for that stream.

EventStreamFactory aids in constructing EventStreams.

CanaryEventProducer creates and POSTs groups of canary events to event services.

Also refactor the EventSchemaLoader JsonLoader classes from refinery-core refinery-core to here.
We will make refinery-source depend on this java library.

NOTE: This code was orignially part of analytics/refinery/source, but has been moved to
this standalone repository.  It was orginally reviewed here:
https://gerrit.wikimedia.org/r/c/analytics/refinery/source/+/603582

Bug: T251609

Change-Id: Id1f57e8ad7edd8f4ad8693024529b1e0a5008829