Remove calls to forbidden APIs.
This was mostly about adding explicit Locales to a few format calls,
and replacing the use of the com.sun HTTP server with Wiremock.

com.sun packages are not part of the Java API and should never be used.
Wiremock introduces a standard HTTP testing framework, which has a
number of facilities to write better HTTP tests (more compact notation,
ability to generate a number of faults, timeouts, etc...).

Bug: T260375
Change-Id: I4dcbe6db698121f2cdc9d08e49ec60160ad4299e