Add functions to EventStreamFactory and EventStreamConfig to match streams by settings
This will allow us to more easily specify settings filters to get back
either only the EventStreams or EventStreamConfigs that we want.

Most immediately, this will be used to target stream topics
for monitoring and ingestion.

Also add toRegex static function to EventStreamConfig,
useful for converting a list of topics to a single regex string.

Bug: T251609
Change-Id: I8d535c3636086b787731a6ae9ad387ec25fe9ef9