Allow using EventStreamConfig without providing an eventServiceToUriMap
This config should probably be moved into the EventStreamConfig API
rather than in this API client.  For now, it is useful to not
have to configure this if a caller is not planning on using
the eventServiceToUriMap (not producing to an event service).

If eventServiceToUriMap is not set, an exception will be thrown
when a method attempts to look up an event service uri, just
as it would if an attemp was made to look up a non existent event service.

Change-Id: I60735121dab57c476a7fedf4983dbe45189ebd0f