[WIP] Introduce pitest to the build.
Just an experiment to see if pitest would make some kind of sense for
simple projects.

This configures the pitest plugin, but does not bind it to a build
phase. See below on how to test locally.

This also adds junit-jupiter-engine as a test dependency. This
dependency is required to run tests. It is automagically added by your
IDE or by the maven surefire plugin, but needs to be explicitely added
for pitest (and since it is a real dependency, it make sense to add
it).

How to test localy:

        ./mvnw org.pitest:pitest-maven:mutationCoverage

The reports are available in
${project}/eventutilities/target/pit-reports/

Potential issues / improvements
* This is adding 30 seconds to the build time.
* The build does not fail if mutation coverage is decreased or below
  some threshold.

Checking a few of the surviving mutations, it seems to be mostly about
error checking. There are a few cases that might be interesting (for
example, the builder in BasicHttpClient seems to lack some tests).

Change-Id: I6b053a7f08741ef92b0695f3c918df5647b69dfa