BasicHttpClient needs to close the underlying HttpClient.
The CloseableHttpClient used by BasicHttpClient is closeable, as
its name indicates. Thus it must be closed. This CR makes
BasicHttpClient implement Closeable.

There are still a few inconsistencies in how resources are
released. For example ResourceLoader can be configured with an
external BasicHttpClient (which it probably should NOT close)
or with an internally created BasicHttpClient (which it should
probably close).

Change-Id: I914e6b9ebcd1df6f5cdb3dd6df5127fb1998080f