Support getting EventStreamConfig settings by JsonPointer path
When getting or filtering EventStreamConfig, if a settingName
starts with a '/', assume it is a JsonPointer path, and use
JsonNode.at() to look up the value rather than JsonNode.get().

This allows for filtering on nested settings.

Also resolve SonarQube warnings in TestEventStream and
TestEventStreamConfig by removing undeeded public modifiers
on @Test functions.

Bug: T273901
Bug: T271232
Change-Id: Ide464dffa861c63230f866d08507f04db475635f