Add withKafkaTimestampAsWatermark as separate flink table builder method
This removes the 'withKafkaTimestampAsWatermark' param on the
setupKafka method, and instead expects you to call it
on its own if you want the kafka timestamp as a watermark.

This allows for easier fine tuning of the watermark delay.
With no params, the default will be to add the kafka timestamp
as a "kafka_timestamp" field with a watermark delay of 10 seconds.

Bug: T308356
Change-Id: I6183f53d60d212db5963757d9ac20c3d54c827ae