Simplify `@Nonnull` annotations by introducing class level defaults.
By adding `@ParametersAreNonnullByDefault` at the class level, the
number of `@Nonnull` annotations can be greatly reduced, while
keeping the same semantics.

Note that in reality the semantics are slightly altered since now all
parameters are considered non null. As far as I can see, this was the
intention all along.

Change-Id: If129e0a4a8a9fb33d0f2013d19d13b7a2e887a45