Add JsonRowSerializationSchema and helper methods to create a KafkaSink
JsonRowSerializationSchema is a class copied & adapted from the flink
code base, it generates a json event following the WMF event platform
rules.

KafkaEventSerializationSchema is a kafka record serialized
wrapping a JsonRowSerializationSchema and is reponsible for setting
the kafka timestamp to the same value as the meta.dt field.

EventRowTypeInfo provides some utility methods to manipulate a Row
object following the WMF event platform rules:

- creates emtpy Row objects supported position and named based access
- project a Row to a target typeInfo possibly intersecting common
  fields and ignoring the fields that must be changed in any new
  events ($schema, meta.id, meta.dt, meta.stream)

It extends RowTypeInfo and is the type returned and used by the
various components built from EventDataStreamFactory.

Also provide an helper function to create a KafkaSinkBuilder based
on a stream name, a schema version and a topic with an optional
partioner.

Bug: T310218
Change-Id: I85cbbf7704e50957a5f41280b89fe49889c08387